Added API for nested overflow:scroll layers, and enable the feature.
The API, WKCoordinatedSceneGetAncestorScrollableContentsLayer, accepts
a graphics layer as its parameter, and returns the nearest ancestor layer
that is also scrollable.

This allows us to nest scrollable layers, e.g. a horizontal and a vertical
layer.

Added API test OverflowScroll that tests the feature, and also enabled it
in the build and added enablers to get the WKCoordinatedScene from WKView.

Reviewed by Marcelo Lira