WTR: Fix crash when DUMPRENDERTREE_TEMP is not set
When running WebKitTestRunner ``by hand'' we may experience a crash due to
an attempt to close the icon database in the destructor of TestController.

Such a crash happens because neither the DUMPRENDERTREE_TEMP env variable
is set nor a valid directory is returned by platformLibraryPathForTesting(),
in TestControllerNix.cpp, which causes the icon database not to be open at
the time the destructor tries to close it.

We don't see this problem when running the ``run-webkit-tests'' script
because DUMPRENDERTREE_TEMP is set by the magical python scripts that
set up the tests -- Tools/Scripts/webkitpy/port/driver.py, in this case.

To verify the crash, simply run something like:
./WebKitBuild/Debug/bin/WebKitTestRunner --help

This patch simply implements the platformLibraryPathForTesting() method,
making it create and return a temporary directory. If DUMPRENDERTREE_TEMP
is set, we don't actually call this method.

Reviewed by: Marcelo Lira