Making WebViewNix scale its content properly
Now the Nix WebView is smart enough so it can find the proper
scale for its content when a page starts loading.
If the comming page is a result of a back/forward command, it just
keeps the scale untouched, letting the cached scale to be applied.
However, if the load is a standard load and fixed layout is being used,
it tries to apply a scale, in order to fit the contents properly

Reviewed by: Daker Fernandes Pinheiro