Fixing TestWebKitAPI CMakeLists file
Leaving it the same as in webkit trunk, so we can avoid conflicts.
The only difference is the TestWebKitAPI_LIBRARIES declaration, that
was moved before the WEBKIT_INCLUDE_CONFIG_FILES_IF_EXISTS so we can
append the CAIRO_LIBRARIES to it.

Also fixed PlatformNix.cmake to look the same as in other ports

Reviewed by: Hugo Parente Lima